+++ src/SpringValue.spec.lua
@@ -1,19 +1,19 @@
 local ReplicatedStorage = game:GetService("ReplicatedStorage")
 
-local RoactSpring = require(ReplicatedStorage.Packages.RoactSpring)
+local SpringValue = require(ReplicatedStorage.Packages.RoactSpring.SpringValue)
 
 return function()
 
     describe("SpringValue", function()
         it("can animate a number", function()
-            local spring = RoactSpring.SpringValue.new({ from =  0 })
+            local spring = SpringValue.new({ from =  0 })
             spring:start({ to = 100 }):await()
             expect(spring.animation:getValue()).to.equal(100)
         end)
 
         describe("when `immediate` prop is true", function()
             it("should stop animating", function()
-                local spring = RoactSpring.SpringValue.new({ from =  0 })
+                local spring = SpringValue.new({ from =  0 })
                 spring:start({ to = 100 })
 
                 task.wait(0.1)
@@ -29,7 +29,7 @@ return function()
 
         describe("the `config` prop", function()
             it("should reset velocity when `to` changes", function()
-                local spring = RoactSpring.SpringValue.new({ from =  0 })
+                local spring = SpringValue.new({ from =  0 })
                 spring:start({ to = 100, config = { velocity = 10 } })
                 expect(spring.animation.config.velocity).to.equal(10)